+++ src/main/java/com/parse/gwt/client/Parse.java
@@ -382,36 +382,45 @@ public class Parse {
                     });
         }
         public static void update(final ParseObject ref, final AsyncCallback<ParseResponse> callback) {
-            String objectId = ref.getObjectId();
-            final String className = ref.getClassName();
-            final String path = Parse.SERVER_URL + Parse.CLASSES_URI + className + "/" + objectId;
-            JSONObject payload = new JSONObject();
-            Iterator<String> it = ref.keySet().iterator();
-            while(it.hasNext()) {
-                String key = it.next();
-                JSONValue value = ref.get(key);
-                // Do not copy objectId
-                if(!key.equals("objectId")){
-                    payload.put(key, value);
+            try {
+                String objectId = ref.getObjectId();
+                final String className = ref.getClassName();
+                final String path = Parse.SERVER_URL + Parse.CLASSES_URI + className + "/" + objectId;
+                JSONObject payload = new JSONObject();
+                Iterator<String> it = ref.keySet().iterator();
+                while(it.hasNext()) {
+                    String key = it.next();
+                    JSONValue value = ref.get(key);
+                    // Do not copy objectId
+                    if(!key.equals("objectId")){
+                        if(key.equals("createdAt") || key.equals("updatedAt")) {
+                            // skip
+                        } else {
+                            payload.put(key, value);
+                        }
+                    }
                 }
+                Shape.put(path)
+                        .header("X-Parse-Application-Id", X_Parse_Application_Id)
+                        .header("X-Parse-REST-API-Key", X_Parse_REST_API_Key)
+                        .header("X-Parse-Master-Key", X_Parse_Master_Key)
+                        .header("X-Parse-Session-Token", X_Parse_Session_Token)
+                        .body(payload.toString())
+                        .asJson(new AsyncCallback<String>() {
+                            @Override
+                            public void onFailure(Throwable throwable) {
+                                HttpRequestException ex = (HttpRequestException) throwable;
+                                callback.onFailure(ex);
+                            }
+                            @Override
+                            public void onSuccess(String s) {
+                                callback.onSuccess(ParseResponse.parse(s));
+                            }
+                        });
+
+            } catch (Exception e) {
+                callback.onFailure(e);
             }
-            Shape.put(path)
-                    .header("X-Parse-Application-Id", X_Parse_Application_Id)
-                    .header("X-Parse-REST-API-Key", X_Parse_REST_API_Key)
-                    .header("X-Parse-Master-Key", X_Parse_Master_Key)
-                    .header("X-Parse-Session-Token", X_Parse_Session_Token)
-                    .body(payload.toString())
-                    .asJson(new AsyncCallback<String>() {
-                        @Override
-                        public void onFailure(Throwable throwable) {
-                            HttpRequestException ex = (HttpRequestException) throwable;
-                            callback.onFailure(ex);
-                        }
-                        @Override
-                        public void onSuccess(String s) {
-                            callback.onSuccess(ParseResponse.parse(s));
-                        }
-                    });
         }
         public static void query(final Query query, final AsyncCallback<ParseResponse> callback) {
             String className = query.get("className").isString().stringValue();