+++ src/main/java/com/parse/gwt/client/autobean/ParseFile.java
@@ -2,9 +2,9 @@ package com.parse.gwt.client.autobean;
 
 public interface ParseFile {
     String get__type();
-    String getClassName();
-    String getObjectId();
+    String getName();
+    String getUrl();
     void set__type(String __type);
-    void getClassName(String className);
-    void setObjectId(String objectId);
+    void setUrl(String url);
+    void setName(String name);
 }