+++ src/main/java/com/parse/gwt/client/js/JsForEachCallbackFn.java
@@ -1,0 +1,8 @@
+package com.parse.gwt.client.js;
+
+import jsinterop.annotations.JsFunction;
+
+@JsFunction
+public interface JsForEachCallbackFn {
+    void onKey(String key);
+}